# api/openapi.yaml
openapi: 3.0.3
info:
  title: happyDomain API
  description: Finally a simple interface for domain names.
  version: "0.1.0"
  contact:
    email: contact+api@happydomain.org
    name: happyDomain team
  license:
    name: GNU Affero General Public License v3.0 or later
    url: https://spdx.org/licenses/AGPL-3.0-or-later.html
...
servers:

  - url: http://localhost:8081/api
    description: Development server

  - url: https://app.happydomain.org/api
    description: Main production server
paths:
...
  /api/users:

    patch:
      operationId: user_special_action
      description: This will send an email to the user either to recover its account
        or with a new email validation link.
      requestBody:
        description: Description of the action to perform and email of the user
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas.yaml#/components/schemas/UserSpecialAction'
      responses:
        "200":
          description: Perhaps something happen
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      summary: Account recovery.
      tags:
        - users

    post:
      operationId: register_user
      description: Register a new happyDomain account (when using internal authentication
        system).
      requestBody:
        description: Account information
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas.yaml#/components/schemas/UserRegistration'
      responses:
        "200":
          description: The created user
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/User'
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      summary: Register account.
      tags:
        - users
...
  /api/users/{userId}/email:

    post:
      operationId: validate_user_email
      description: This is the route called by the web interface in order to validate
        the e-mail address of the user.
      parameters:
      - description: User identifier
        in: path
        name: userId
        required: true
        schema:
          type: string
      requestBody:
        description: Validation form
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas.yaml#/components/schemas/AddressValidationForm'
      responses:
        "204":
          description: Email validated, you can now login
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      summary: Validate e-mail address.
      tags:
        - users
...
  /api/users/{userId}/recovery:

    post:
      operationId: recover_user_account
      description: "This performs account recovery by reseting the password of
        the account."
      parameters:
      - description: User identifier
        in: path
        name: userId
        required: true
        schema:
          type: string
      requestBody:
        description: Recovery form
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas.yaml#/components/schemas/AccountRecoveryForm'
      responses:
        "204":
          description: Recovery completed, you can now login with your new credentials
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      summary: Reset password with link in email.
      tags:
        - users
...
components:
  securitySchemes:

    ApiKeyAuth:
      description: Description for what is this security definition being used
      in: header
      name: Authorization
      type: apiKey

    BasicAuth:
      type: http
      scheme: basic
...